Added data_simulation_utils.py and moved util functions
